re-establish typescript
