+++ src/config.ts
@@ -1,0 +1,22 @@
+import { getTimestamp } from "./util/timestamp";
+import BuildInfo from "../config.json";
+
+export default function (eleventyConfig: any) {
+  eleventyConfig.addPassthroughCopy({
+    "./src/static/": "/",
+  });
+  eleventyConfig.addGlobalData("buildTimeEncoded", () => {
+    return getTimestamp(Number(BuildInfo.version));
+  });
+  eleventyConfig.addWatchTarget("./src/config.ts");
+
+  return {
+    templateFormats: ["njk", "html"],
+    htmlTemplateEngine: "njk",
+    dir: {
+      input: "src/pages",
+      includes: "../templates",
+      output: "public",
+    },
+  };
+}
